feat:管理员添加 删除 修改商户

+++ src/main/java/com/example/pj/entity/Merchant.java
@@ -14,7 +14,7 @@ import org.springframework.stereotype.Component;
 
 public class Merchant {
 
-    private Integer merchantId;
+    private Long merchantId;
 
     private String merchantName;
